A couple of fixes for path handling
It appears that indexing into the files in a git tree only works with
POSIX paths, so the Windows native paths need to be converted.

Additionally, I made a bad assumption about the 'python.yaml' file name
in the rosdep analyzer. That change should have no effect on POSIX
systems, and makes Windows runs more consistent.