Do not register extension points with colcon
As an independent tool built on colcon's framework, this tool should not
register its extension points as colcon extension points. This should
prevent it from appearing in colcon-devtools output, namely
version-check.

This isn't to say that extension point registration and presence in
colcon-devtools wouldn't provide any benefit, this is an effort to avoid
"kitchen sinking" colcon as a command line tool.