Add special logic for handling 'key-ordering' in yamllint
The key-ordering linter needs special consideration. It scans from top
to bottom, so when a key is added (or modified) which should be earlier,
it is identified properly. However, if a key is added which should be
placed later, the key AFTER it is flagged as the first key which is out
of order.

When we see an out-of-order problem that isn't on a modified line, check
to see if the key before it was changed. If so, it is almost certainly
the real cause.