Don't complain about version suffixes in deb names (#47)
There is a rosdep key naming guideline that key names should generally
be the same as the Ubuntu deb names that back them. It is not uncommon
to omit any major version suffixes that appear in the package name when
naming the key, as that pattern isn't generally used on other platforms
and increases rosdep database maintenance burden.

This change relaxes the check to allow keys to omit the version number.
This check isn't blocking and is more of a guideline, but it's almost
always noise and viewed as a false-positive when this is encountered.